skills/audio-transcriber-skill/pyproject.toml
[project]
name = "audio-transcriber-skill"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["openai-whisper", "torch"]
